+++ src/nanodsp/__init__.py
@@ -2,6 +2,6 @@
 
 from nanodsp.buffer import AudioBuffer
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"
 
 __all__ = ["AudioBuffer", "__version__"]